[ghstack-poisoned]

+++ thunder/core/dtypes.py
@@ -599,8 +599,8 @@ def to_torch_dtype(x: None | torch.dtype | dtype) -> None | torch.dtype:
 _thunder_to_numpy_dtype_map = {
     bool: np.bool_,
     int: np.int_,
-    float: np.float_,
-    complex: np.cfloat,
+    float: np.float64,
+    complex: np.complex128,
     bool8_: np.bool_,
     bool8: np.bool_,
     uint8_: np.uint8,
